+++ lib/cgraph/refstr.c
@@ -14,9 +14,10 @@
  *************************************************************************/
 
 #include <cgraph/cghdr.h>
-#include <stddef.h>
 #include <stdbool.h>
 #include <stdio.h>
+#include <stdlib.h>
+#include <util/alloc.h>
 
 /*
  * reference counted strings.
@@ -35,7 +36,7 @@ static Dtdisc_t Refstrdisc = {
     -1,				/* size */
     0,				/* link offset */
     NULL,
-    agdictobjfree,
+    free,
     NULL,
 };
 
@@ -111,7 +112,7 @@ static char *agstrdup_internal(Agraph_t *g, const char *s, bool is_html) {
     else {
 	sz = sizeof(refstr_t) + strlen(s);
 	if (g)
-	    r = agalloc(g, sz);
+	    r = gv_calloc(sz, sizeof(char));
 	else {
 	    r = malloc(sz);
 	    if (sz > 0 && r == NULL) {